requirements.txt
astroid==2.14.2
click==8.1.3
dill==0.3.6
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
mysql-connector==2.2.9
mysql-connector-python==8.0.32
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
protobuf==3.20.3
pylint==2.16.2
PyMySQL==1.0.2
SQLAlchemy==2.0.4
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.5.0
Werkzeug==2.2.3
wrapt==1.14.1
